ml_model/requirements.txt
# Updated requirements with compatible versions
Flask>=2.3.3
flask-cors>=4.0.0
numpy>=1.23.5
gunicorn>=21.2.0
h5py>=3.7.0
scikit-learn>=1.1.3
protobuf>=3.19.6
# Environment variables
python-dotenv>=1.0.0
# Firebase Admin SDK
firebase-admin>=6.0.0
google-cloud-firestore>=2.0.0
# TensorFlow is optional - uncomment if needed
# tensorflow>=2.20.0